refactor: refactor cards component
