+++ react/src/components/AboutUs/AboutUs.tsx
@@ -1,7 +1,3 @@
-import { Component } from 'react';
-
-export class AboutUs extends Component {
-  render() {
-    return <main>About us</main>;
-  }
-}
+export const AboutUs = () => {
+  return <main>About us</main>;
+};
